+++ src/app/control/dev-home/dialogs/graph-content-connection/dialog-graph-connection-view.component.ts
