+++ src/app/control/dev-home/graph/graph.component.scss
