+++ src/app/control/navbar/dialogs/add-appllication/dialog-add-application.html
