+++ src/app/control/navbar/dialogs/add-cluster/dialog-add-cluster.scss
