+++ src/app/control/navbar/dialogs/generate-token/dialog-generate-token.scss
