+++ src/app/control/profile/profile.component.scss
