+++ src/app/control/users/users.component.ts
@@ -6,9 +6,9 @@ import { DatePipe } from '@angular/common';
 import { select, Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import { ApiService } from '../../shared/modules/api/api.service';
-import { DialogEditUserView } from '../dialogs/edit-user/dialogEditUser';
+import { DialogEditUserView } from './dialogs/edit-user/dialogEditUser';
 import { NotificationService } from '../../shared/modules/notification/notification.service';
-import { DialogConfirmationView } from '../dialogs/confirmation/dialogConfirmation';
+import { DialogConfirmationView } from '../../root/components/dialogs/confirmation/dialogConfirmation';
 import { IUser } from '../../root/interfaces/user';
 import { DialogAction } from '../../root/enums/dialogAction';
 import { IDialogAttribute } from '../../root/interfaces/dialogAttribute';
