+++ src/app/root/components/dialogs/confirmation/dialogConfirmation.ts
@@ -1,6 +1,6 @@
 import { Component, Inject } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { DialogGraphConnectionView } from '../graph-content-connection/dialog-graph-connection-view.component';
+import { DialogGraphConnectionView } from '../../../../control/dev-home/dialogs/graph-content-connection/dialog-graph-connection-view.component';
 
 interface ConformationText {
     text: string;
